Refactor 'Workflow Actions' to 'Workflow Operations'
